requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional (for Jupyter notebooks)
jupyter>=1.0.0
ipython>=8.12.0

# Optional (for downloading from Google Drive)
gdown>=4.7.0

# Web app
streamlit>=1.32.0
